refactor: rename File Access class
