+++ BYVAM0/BYVAM0/DataAccess/FileAccessor.cs
@@ -2,10 +2,11 @@
 using System;
 
 using BYVAM0.Model;
+using BYVAM0.Interfaces;
 
 namespace BYVAM0.DataAccess
 {
-    internal class FileAccess : IDataAccess
+    internal class FileAccessor : IDataAccess
     {
         private static readonly JsonSerializerOptions Options = new()
         {
@@ -21,7 +22,7 @@ namespace BYVAM0.DataAccess
 
                 return questions;
             }
-            catch (Exception ex) when (ex is FileNotFoundException 
+            catch (Exception ex) when (ex is FileNotFoundException
                                        || ex is DirectoryNotFoundException)
             {
                 return null;
