+++ BYVAM0/BYVAM0/Interfaces/IDataAccess.cs
@@ -1,6 +1,6 @@
 ﻿using BYVAM0.Model;
 
-namespace BYVAM0.DataAccess
+namespace BYVAM0.Interfaces
 {
     internal interface IDataAccess
     {
